# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Spring Swagger Integration",
    "description": "This page lists all the rest apis for Spring REST Application.",
    "version": "1.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "//localhost:8080/"
    }
  ],
  "tags": [
    {
      "name": "customer-controller",
      "description": "Customer Controller"
    }
  ],
  "paths": {
    "/customer": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "summary": "getCustomers",
        "operationId": "getCustomersUsingGET",
        "parameters": [
          {
            "name": "first",
            "in": "query",
            "description": "first",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "last",
            "in": "query",
            "description": "last",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "type",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "customer-controller"
        ],
        "summary": "postCustomers",
        "operationId": "postCustomersUsingPOST",
        "requestBody": {
          "description": "pCustomer",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    },
    "/customer/{id}": {
      "get": {
        "tags": [
          "customer-controller"
        ],
        "summary": "getCustomerById",
        "operationId": "getCustomerByIdUsingGET",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "customer-controller"
        ],
        "summary": "putCustomers",
        "operationId": "putCustomersUsingPUT",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "pCustomer",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "customer-controller"
        ],
        "summary": "deleteCustomer",
        "operationId": "deleteCustomerUsingDELETE",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {}
          }
        }
      },
      "patch": {
        "tags": [
          "customer-controller"
        ],
        "summary": "patchCustomer",
        "operationId": "patchCustomerUsingPATCH",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "patched",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "dob": {
            "type": "string",
            "format": "date"
          },
          "first": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "last": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PRIORITY",
              "NORMAL"
            ]
          }
        }
      }
    }
  }
}